fix(newrelic): Attach userId and chatId to error inbox events
NewRelic Errors Inbox showed an empty Users column because nothing
called setUserID per request. Add a Fastify preHandler that sets
enduser.id and tg.* custom attributes from the Telegram webhook body,
and unify the body parsing with Sentry behind a shared hook helper.
